move Exception and ConfigHelper to utilities folder
