+++ build.gradle
@@ -11,7 +11,7 @@ apply plugin: 'net.minecraftforge.gradle.forge'
 //Only edit below this line, the above code adds and enables the necessary things for Forge to be setup.
 
 
-version = "0.5"
+version = "0.6"
 group = "com.garthi.timebanmod" // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = "timebanmod"
